src/requirements.txt
pytest
SpeechRecognition
PyAudio
google-generativeai
pyttsx3
python-dotenv